Replace global slot objects for the convenience alpha functions by
* clutter/src/alpha.{ccg,hg} (ALPHA_*): Replace global slot objects
for the convenience alpha functions by ordinary static methods of
class Alpha.  This is just as easy to use but simpler and without
the trouble associated with the runtime initialization  of global
objects in dynamic libraries.  Also adapt the code to the additions
and syntax changes in the Clutter API.
(Alpha::Alpha): Remove the ctor which takes a plain ClutterFunction
argument.  If absolutely necessary, there is always the C API.

svn path=/cluttermm/trunk/; revision=1877