Rename to noinst_PROGRAMS, or every example executable will be installed
* Makefile.am (bin_PROGRAMS): Rename to noinst_PROGRAMS, or every
example executable will be installed as $(bindir)/example -- the
last one wins.  Do we want to install the example binaries, and if
so, where?
({,un}install-tutorial): Depend on docs/tutorial/html/index.html
rather than the phony target html, because the rule derives the
location of the html build directory from the first dependency.

svn path=/cluttermm; revision=1953